#899 refactor(pages/deck-start): move DeckStart filter UI and state from features/study
