+++ src/pages/card-list/ui/CardListPage.spec.tsx
@@ -55,14 +55,16 @@ vi.mock("@/features/card-list", () => ({
   },
 }));
 vi.mock("@/features/study", () => ({
+  useEditStudyProgress: () => ({ updateBy: mocks.updateBy }),
+  useStudyCards: (deck: Deck | undefined, cards: Card[]) => (deck == null ? [] : cards),
+}));
+vi.mock("@/pages/deck-start", () => ({
   DeckStartForm: () => <div>Filter controls</div>,
   useDeckFilterState: () => ({
     scoreMax: 4,
     scoreMin: -2,
     tagFilterProps: { selectedTags: ["typescript"], onClickTag: mocks.onClickTag },
   }),
-  useEditStudyProgress: () => ({ updateBy: mocks.updateBy }),
-  useStudyCards: (deck: Deck | undefined, cards: Card[]) => (deck == null ? [] : cards),
 }));
 vi.mock("react-router-dom", () => ({
   useParams: () => mocks.params,
